requirements.txt
mistralai
python-dotenv
tqdm
tenacity
openai
faiss-cpu
sentence-transformers
numpy
